orderbook refactoring
