+++ package.json
@@ -56,6 +56,7 @@
 		"babel-plugin-styled-components": "^1.10.6",
 		"babel-preset-react-app": "~6.1.0",
 		"bfj": "6.1.1",
+		"browserslist": "^4.8.3",
 		"case-sensitive-paths-webpack-plugin": "2.1.2",
 		"chalk": "2.4.1",
 		"css-loader": "~1.0.1",
@@ -97,12 +98,12 @@
 		"stylelint-config-recommended": "^3.0.0",
 		"stylelint-config-styled-components": "^0.1.1",
 		"stylelint-processor-styled-components": "^1.9.0",
+		"uglifyjs-webpack-plugin": "2.2.0",
 		"url-loader": "1.1.1",
 		"webpack": "4.19.1",
 		"webpack-dev-server": "3.8.0",
 		"webpack-manifest-plugin": "2.0.4",
-		"workbox-webpack-plugin": "3.6.3",
-		"uglifyjs-webpack-plugin": "2.2.0"
+		"workbox-webpack-plugin": "3.6.3"
 	},
 	"jest": {
 		"collectCoverageFrom": [
