+++ src/components/OrderBook/OrderBook.js
@@ -1,44 +1,25 @@
 import React, { useState, useEffect } from 'react';
 import { connect } from 'react-redux';
 import styled, { withTheme } from 'styled-components';
-import { format } from 'date-fns';
 
 import { DataSmall } from '../Typography';
-import { Table, Tr, Thead, Tbody, Th, Td, DataLabel } from '../Table';
+import CurrentOrders from './CurrentOrders';
 
 import { getTransactions, getPendingTransactions, getWalletInfo } from '../../ducks/';
 import { updateTransaction, removePendingTransaction } from '../../ducks/transaction';
 
 import snxJSConnector from '../../utils/snxJSConnector';
+import PastTransactions from './PastTransactions';
 
-const StatusLabel = ({ transaction: { status, hash }, colors, network, hasHash }) => {
-	let color = '';
-	switch (status) {
-		case 'Cancelled':
-		case 'Failed':
-			color = colors.red;
-			break;
-		case 'Confirmed':
-			color = colors.green;
-			break;
-		case 'Pending':
-			color = colors.fontTertiary;
-			break;
-		default:
-			color = '';
-			break;
+const BookContent = ({ tab, transactions }) => {
+	switch (tab) {
+		case 'Your orders':
+			return <CurrentOrders transactions={transactions} />;
+		case 'Your trades':
+			return <PastTransactions />;
+		case 'All trades':
+			return <PastTransactions />;
 	}
-	return hasHash ? (
-		<Link
-			color={color}
-			href={`https://${network === 'mainnet' ? '' : network + '.'}etherscan.io/tx/${hash}`}
-			target="_blank"
-		>
-			<DataLabel color={color}>{status}</DataLabel>
-		</Link>
-	) : (
-		<DataLabel color={color}>{status}</DataLabel>
-	);
 };
 
 const OrderBook = ({
@@ -47,7 +28,6 @@ const OrderBook = ({
 	pendingTransactions,
 	updateTransaction,
 	removePendingTransaction,
-	walletInfo: { networkName },
 }) => {
 	const [activeTab, setActiveTab] = useState('Your orders');
 
@@ -78,13 +58,11 @@ const OrderBook = ({
 	return (
 		<Container>
 			<Tabs>
-				{['Your orders', 'Your trades', 'Show all trades'].map(tab => {
+				{['Your orders', 'Your trades', 'All trades'].map(tab => {
 					return (
 						<Tab
-							isDisabled={['Your trades', 'Show all trades'].includes(tab)}
 							key={tab}
 							onClick={() => {
-								if (['Your trades', 'Show all trades'].includes(tab)) return;
 								setActiveTab(tab);
 							}}
 							hidden={!tab}
@@ -98,57 +76,7 @@ const OrderBook = ({
 				})}
 			</Tabs>
 			<Book>
-				<Table cellSpacing="0">
-					<Thead>
-						<Tr>
-							{['Date | Time', 'Pair', 'Price', 'Amount', 'Total', 'Status'].map((label, i) => {
-								return (
-									<Th key={i}>
-										<ButtonSort>
-											<DataSmall color={colors.fontTertiary}>{label}</DataSmall>
-											{/* <SortIcon src={'/images/sort-arrows.svg'} /> */}
-										</ButtonSort>
-									</Th>
-								);
-							})}
-						</Tr>
-					</Thead>
-					<Tbody>
-						{transactions.map((transaction, i) => {
-							const { date, pair, price, amount, totalUSD, status } = transaction;
-							const isPending = ['Pending', 'Waiting'].includes(status);
-							const hasHash = !!transaction.hash;
-							return (
-								<TransactionRow isPending={isPending} key={i}>
-									<Td>
-										<DataLabel>{format(date, 'DD-MM-YY')}</DataLabel>
-									</Td>
-									<Td>
-										<DataLabel>{pair}</DataLabel>
-									</Td>
-									<Td>
-										<DataLabel>{price}</DataLabel>
-									</Td>
-									<Td>
-										<DataLabel>{amount}</DataLabel>
-									</Td>
-									<Td>
-										<DataLabel>${totalUSD}</DataLabel>
-									</Td>
-									<Td>
-										<StatusLabel
-											transaction={transaction}
-											colors={colors}
-											network={networkName}
-											hasHash={hasHash}
-										/>
-									</Td>
-								</TransactionRow>
-							);
-						})}
-						<Tr></Tr>
-					</Tbody>
-				</Table>
+				<BookContent tab={activeTab} transactions={transactions} />
 			</Book>
 		</Container>
 	);
@@ -201,32 +129,6 @@ const Book = styled.div`
 // 	height: 8px;
 // 	margin-left: 5px;
 // `;
-
-const ButtonSort = styled.button`
-	text-align: left;
-	display: flex;
-	align-items: center;
-	border: none;
-	outline: none;
-	cursor: pointer;
-	background-color: transparent;
-	padding: 0;
-`;
-
-const Link = styled.a`
-	cursor: pointer;
-	display: flex;
-	align-items: center;
-	justify-content: flex-end;
-	text-decoration: underline;
-	color: ${props => props.color};
-`;
-
-const TransactionRow = styled(Tr)`
-	& span {
-		color: ${props => (props.isPending ? props.theme.colors.fontTertiary : '')};
-	}
-`;
 
 const mapStateToProps = state => {
 	return {
