+++ src/components/OrderBook/PastTransactions.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const PastTransaction = () => {
+	return <div>here</div>;
+};
+
+export default PastTransaction;
